added parallel execution for tests deleted sample tests and added just two tests in different story files refactoring according to jbehave archetype
