+++ src/test/java/com/gmail/shnapi007/Acceptance.java
@@ -1,0 +1,11 @@
+package com.gmail.shnapi007;
+
+import net.serenitybdd.jbehave.SerenityStories;
+
+
+public class Acceptance extends SerenityStories {
+
+    public Acceptance() {
+        findStoriesCalled("acceptance.story");
+    }
+}
