+++ src/test/java/com/gmail/shnapi007/Another.java
@@ -1,0 +1,10 @@
+package com.gmail.shnapi007;
+
+import net.serenitybdd.jbehave.SerenityStories;
+
+
+public class Another extends SerenityStories {
+    public Another() {
+        findStoriesCalled("another.story");
+    }
+}
