--- src/test/java/com/gmail/shnapi007/TestRunner.java
@@ -1,7 +1,0 @@
-package com.gmail.shnapi007;
-
-import net.serenitybdd.jbehave.SerenityStories;
-
-public class TestRunner extends SerenityStories {
-
-}
