+++ src/test/resources/serenity.conf
@@ -10,7 +10,6 @@ drivers {
 serenity {
   project.name = "sernity-jbehave-selenium"
   take.screenshots = FOR_FAILURES
-  test.root = "com.gmail.shnapi007.steps"
 }
 
 junit.retry.tests = true
