+++ src/test/resources/stories/acceptance/acceptance.story
@@ -1,0 +1,5 @@
+
+Scenario: User searches for picture on Google
+Given Accesses the google page
+When When they type in picture
+Then We should see picture
