+++ src/test/resources/stories/another/another.story
@@ -1,0 +1,4 @@
+
+Scenario: User open DuckDuckGo
+Given Accesses the DuckDuckGo page
+Then We should see DuckDuckGo in title
